+++ Editor/Dropdown/SerializeReferencePropertyDrawer.cs
@@ -3,28 +3,21 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.Serialization;
-using SerializeReferenceDropdown.Editor.Analyzer;
 using SerializeReferenceDropdown.Editor.Preferences;
 using SerializeReferenceDropdown.Editor.Utils;
 using UnityEditor;
-using UnityEditor.IMGUI.Controls;
-using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.Serialization;
-using UnityEngine.UIElements;
 using Object = UnityEngine.Object;
 
 namespace SerializeReferenceDropdown.Editor.Dropdown
 {
     [CustomPropertyDrawer(typeof(SerializeReferenceDropdownAttribute))]
-    public class SerializeReferencePropertyDrawer : PropertyDrawer
+    public partial class SerializeReferencePropertyDrawer : PropertyDrawer
     {
         private const string NullName = "null";
         private List<Type> assignableTypes;
         private Rect propertyRect;
-
-        //TODO Need find better solution for check ui update and traverse all serialized properties
-        private static bool isDirtyUIToolkit;
 
         //TODO: need to find better solution
         private static readonly Dictionary<Object, HashSet<string>> targetObjectAndSerializeReferencePaths =
@@ -39,191 +32,6 @@ namespace SerializeReferenceDropdown.Editor.Dropdown
             var index = Array.FindIndex(refsArray, t => t == refId);
             var hue = (float)index / refsArray.Length;
             return Color.HSVToRGB(hue, 0.8f, 0.8f);
-        }
-
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return EditorGUI.GetPropertyHeight(property, label, true);
-        }
-
-        public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
-        {
-            EditorGUI.BeginProperty(rect, label, property);
-
-            propertyRect = rect;
-
-            if (property.propertyType == SerializedPropertyType.ManagedReference)
-            {
-                DrawIMGUITypeDropdown(rect, property, label);
-            }
-            else
-            {
-                EditorGUI.PropertyField(rect, property, label, true);
-            }
-
-            EditorGUI.EndProperty();
-        }
-
-        public override VisualElement CreatePropertyGUI(SerializedProperty property)
-        {
-            var root = new VisualElement();
-            if (property.propertyType == SerializedPropertyType.ManagedReference)
-            {
-                DrawUIToolkitTypeDropdown(root, property);
-            }
-            else
-            {
-                root.Add(new PropertyField(property));
-            }
-
-            return root;
-        }
-
-        private void DrawUIToolkitTypeDropdown(VisualElement root, SerializedProperty property)
-        {
-            var hideStyle = new StyleEnum<DisplayStyle>(DisplayStyle.None);
-            var flexStyle = new StyleEnum<DisplayStyle>(DisplayStyle.Flex);
-            bool isNew = true;
-            var uiToolkitLayoutPath =
-                "Packages/com.alexeytaranov.serializereferencedropdown/Editor/Layouts/SerializeReferenceDropdown.uxml";
-            var visualTreeAsset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(uiToolkitLayoutPath);
-            root.Add(visualTreeAsset.Instantiate());
-            var propertyField = root.Q<PropertyField>();
-
-            var selectTypeButton = root.Q<Button>("typeSelect");
-            selectTypeButton.clickable.clicked += ShowDropdown;
-
-            var fixCrossRefButton = root.Q<Button>("fixCrossReferences");
-            fixCrossRefButton.clickable.clicked += () =>
-            {
-                MakeDirtyUIToolkit();
-                FixCrossReference(property);
-            };
-            var openSourceFIleButton = root.Q<Button>("openSourceFile");
-            openSourceFIleButton.style.display = hideStyle;
-            openSourceFIleButton.clicked += () => { OpenSourceFile(property.managedReferenceValue.GetType()); };
-            if (SerializeReferenceToolsUserPreferences.GetOrLoadSettings().DisableOpenSourceFile == false)
-            {
-                openSourceFIleButton.style.display = property.managedReferenceValue == null ? hideStyle : flexStyle;
-            }
-
-            var propertyPath = property.propertyPath;
-            assignableTypes ??= GetAssignableTypes(property);
-            root.TrackSerializedObjectValue(property.serializedObject, UpdateDropdown);
-            UpdateDropdown(property.serializedObject);
-            isNew = false;
-
-            void ShowDropdown()
-            {
-                var dropdown = new SerializeReferenceAdvancedDropdown(new AdvancedDropdownState(),
-                    assignableTypes.Select(GetTypeName), index =>
-                    {
-                        MakeDirtyUIToolkit();
-                        WriteNewInstanceByIndexType(index, property);
-                    });
-                var buttonMatrix = selectTypeButton.worldTransform;
-                var position = new Vector3(buttonMatrix.m03, buttonMatrix.m13, buttonMatrix.m23);
-                var buttonRect = new Rect(position, selectTypeButton.contentRect.size);
-                dropdown.Show(buttonRect);
-            }
-
-            void UpdateDropdown(SerializedObject so)
-            {
-                var prop = so.FindProperty(propertyPath);
-                propertyField.BindProperty(prop);
-                var selectedType = TypeUtils.ExtractTypeFromString(prop.managedReferenceFullTypename);
-                var selectedTypeName = GetTypeName(selectedType);
-                selectTypeButton.text = selectedTypeName;
-                selectTypeButton.tooltip = $"Class: {selectedType?.Name}\nNamespace: {selectedType?.Namespace}";
-                if (isNew == false && isDirtyUIToolkit == false)
-                {
-                    return;
-                }
-
-                selectTypeButton.style.color = new StyleColor(Color.white);
-                fixCrossRefButton.style.display = hideStyle;
-
-                if (IsHaveSameOtherSerializeReference(property))
-                {
-                    fixCrossRefButton.style.display = flexStyle;
-                    var color = GetColorForEqualSerializedReference(property);
-                    selectTypeButton.style.color = color;
-                }
-            }
-
-            void MakeDirtyUIToolkit()
-            {
-                isDirtyUIToolkit = true;
-                EditorApplication.delayCall += () => { isDirtyUIToolkit = false; };
-            }
-        }
-
-        private void DrawIMGUITypeDropdown(Rect rect, SerializedProperty property, GUIContent label)
-        {
-            const float fixButtonWidth = 40f;
-            assignableTypes ??= GetAssignableTypes(property);
-
-            var isHaveOtherReference = IsHaveSameOtherSerializeReference(property);
-
-            var referenceType = TypeUtils.ExtractTypeFromString(property.managedReferenceFullTypename);
-
-            var dropdownRect = GetDropdownIMGUIRect(rect);
-
-            EditorGUI.EndDisabledGroup();
-
-            var dropdownTypeContent = new GUIContent(
-                text: GetTypeName(referenceType),
-                tooltip: GetTypeTooltip(referenceType));
-
-            var style = EditorStyles.miniPullDown;
-            if (isHaveOtherReference)
-            {
-                var uniqueColor = GetColorForEqualSerializedReference(property);
-                style = new GUIStyle(EditorStyles.miniPullDown)
-                    { normal = new GUIStyleState() { textColor = uniqueColor } };
-            }
-
-            if (EditorGUI.DropdownButton(dropdownRect, dropdownTypeContent, FocusType.Keyboard, style))
-            {
-                var dropdown = new SerializeReferenceAdvancedDropdown(new AdvancedDropdownState(),
-                    assignableTypes.Select(GetTypeName),
-                    index => WriteNewInstanceByIndexType(index, property));
-                dropdown.Show(dropdownRect);
-            }
-
-            if (isHaveOtherReference)
-            {
-                if (GUI.Button(GetFixCrossReferencesRect(dropdownRect), "Fix"))
-                {
-                    FixCrossReference(property);
-                }
-            }
-
-            EditorGUI.PropertyField(rect, property, label, true);
-
-
-            Rect GetDropdownIMGUIRect(Rect mainRect)
-            {
-                var dropdownOffset = EditorGUIUtility.labelWidth;
-                Rect rect = new Rect(mainRect);
-                rect.width -= dropdownOffset;
-                rect.x += dropdownOffset;
-                rect.height = EditorGUIUtility.singleLineHeight;
-                if (isHaveOtherReference)
-                {
-                    rect.width -= fixButtonWidth;
-                }
-
-                return rect;
-            }
-
-            Rect GetFixCrossReferencesRect(Rect rectIn)
-            {
-                var newRect = rectIn;
-                newRect.x += rectIn.width + EditorGUIUtility.standardVerticalSpacing;
-                newRect.width = fixButtonWidth - EditorGUIUtility.standardVerticalSpacing;
-                return newRect;
-            }
         }
 
         private bool IsHaveSameOtherSerializeReference(SerializedProperty property)
@@ -432,6 +240,7 @@ namespace SerializeReferenceDropdown.Editor.Dropdown
             {
                 var needSaveData = SerializeReferenceToolsUserPreferences.GetOrLoadSettings().CopyDataWithNewType;
                 var targets = property.serializedObject.targetObjects;
+                
                 // Multiple object edit.
                 //One Serialized Object for multiple Objects work sometimes incorrectly  
                 foreach (var target in targets)
